+++ src/hash/calcHash.js
@@ -1,5 +1,5 @@
 const calculateHash = async () => {
-    // Write your code here 
+    
 };
 
 await calculateHash();
